♻️ refactor : Lotto 클래스 정리
