+++ src/domain/Lotto.js
@@ -1,12 +1,10 @@
+import { LOTTO_LENGTH, MIN_NUMBER, MAX_NUMBER } from "../utils/constants.js";
 import {
   TYPE_ERROR,
   RANGE_ERROR,
   DUPLICATE_ERROR,
   LOTTO_LENGTH_ERROR,
-  LOTTO_LENGTH,
-  MIN_NUMBER,
-  MAX_NUMBER,
-} from "./utils/constants.js";
+} from "../utils/errorMessage.js";
 
 class Lotto {
   #numbers;
